refactor: Add new util subdirectory in components
