+++ client/src/components/util/Notify.js
@@ -2,7 +2,7 @@ import { Alert, Snackbar, Slide } from '@mui/material';
 import React from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
-import { notifyAction } from '../actions/actions';
+import { notifyAction } from '../../actions/actions';
 
 const Notify = () => {
     const dispatch = useDispatch();
